+++ cpp/open3d/data/CMakeLists.txt
@@ -2,7 +2,7 @@ open3d_ispc_add_library(data OBJECT)
 
 target_sources(data PRIVATE
     Dataset.cpp
-    Downloader.cpp
+    Download.cpp
 )
 
 open3d_show_and_abort_on_warning(data)
