+++ cpp/tests/data/CMakeLists.txt
@@ -1,4 +1,4 @@
 target_sources(tests PRIVATE
     Dataset.cpp
-    Downloader.cpp
+    Download.cpp
 )
